fix(rust): correct the condition for skipping serialization
**Problem:** The generated rust code incorrectly applies
`#[serde(skip_serializing_if = "Option::is_none")]` not when fields are
*optional*, but when they are *nullable*, i.e. their type is `foo |
null`. This effectively does the opposite of what should be done, as it
doesn't serialize the field when in fact it should be serialized and be
given the value `null`. Conversely, optional properties which should not
be serialized at all (if `None`) *are* serialized and given the value of
`null`.

**Solution:** Always skip serializing on optional fields, regardless of
whether they are nullable or not.